# crates/devolutions-pedm-shared/devolutions-pedm-client-http/src/models/elevation_method.rs
/*
 * Devolutions PEDM API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document:
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

///
#[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
pub enum ElevationMethod {
    #[serde(rename = "LocalAdmin")]
    LocalAdmin,
    #[serde(rename = "VirtualAccount")]
    VirtualAccount,
}

impl std::fmt::Display for ElevationMethod {
    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
        match self {
            Self::LocalAdmin => write!(f, "LocalAdmin"),
            Self::VirtualAccount => write!(f, "VirtualAccount"),
        }
    }
}

impl Default for ElevationMethod {
    fn default() -> ElevationMethod {
        Self::LocalAdmin
    }
}