# crates/devolutions-pedm-shared/devolutions-pedm-client-http/src/models/get_profiles_me_response.rs
/*
 * Devolutions PEDM API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document:
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct GetProfilesMeResponse {
    #[serde(rename = "Active", skip_serializing_if = "Option::is_none")]
    pub active: Option<String>,
    #[serde(rename = "Available")]
    pub available: Vec<String>,
}

impl GetProfilesMeResponse {
    pub fn new(available: Vec<String>) -> GetProfilesMeResponse {
        GetProfilesMeResponse {
            active: None,
            available,
        }
    }
}