# crates/devolutions-pedm-shared/devolutions-pedm-client-http/src/models/profile.rs
/*
 * Devolutions PEDM API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document:
 *
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct Profile {
    #[serde(rename = "DefaultElevationKind", skip_serializing_if = "Option::is_none")]
    pub default_elevation_kind: Option<models::ElevationKind>,
    #[serde(rename = "ElevationMethod", skip_serializing_if = "Option::is_none")]
    pub elevation_method: Option<models::ElevationMethod>,
    #[serde(rename = "ElevationSettings", skip_serializing_if = "Option::is_none")]
    pub elevation_settings: Option<models::ElevationConfigurations>,
    #[serde(rename = "Id", skip_serializing_if = "Option::is_none")]
    pub id: Option<String>,
    #[serde(rename = "Name", skip_serializing_if = "Option::is_none")]
    pub name: Option<String>,
    #[serde(rename = "PromptSecureDesktop", skip_serializing_if = "Option::is_none")]
    pub prompt_secure_desktop: Option<bool>,
    #[serde(rename = "Rules", skip_serializing_if = "Option::is_none")]
    pub rules: Option<Vec<String>>,
}

impl Profile {
    pub fn new() -> Profile {
        Profile {
            default_elevation_kind: None,
            elevation_method: None,
            elevation_settings: None,
            id: None,
            name: None,
            prompt_secure_desktop: None,
            rules: None,
        }
    }
}