# crates/devolutions-pedm/openapi/dotnet-client/api/openapi.yaml
openapi: 3.1.0
info:
  title: Devolutions PEDM API
  version: ""
servers:
- url: /
paths:
...
  /policy/me:
    get:
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfilesMeResponse'
          description: ""
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionalId'
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
  /policy/profiles:
    get:
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Id'
                type: array
          description: ""
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
  /policy/profiles/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ""
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Profile'
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
  /policy/rules:
    get:
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Id'
                type: array
          description: ""
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rule'
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
  /policy/rules/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
          description: ""
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rule'
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
  /policy/assignments:
    get:
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Assignment'
                type: array
          description: ""
  /policy/assignments/{id}:
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/User'
              type: array
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ""
        "200":
          description: no content
components:
  schemas:
...
    ApplicationFilter:
      example:
        Path:
          Kind: Equals
          Data: Data
        WorkingDirectory:
          Kind: Equals
          Data: Data
        CommandLine:
        - Kind: Equals
          Data: Data
        - Kind: Equals
          Data: Data
        Hashes:
        - Sha1: Sha1
          Sha256: Sha256
        - Sha1: Sha1
          Sha256: Sha256
        Signature:
          CheckAuthenticode: true
      properties:
        CommandLine:
          items:
            $ref: '#/components/schemas/StringFilter'
          type: array
          nullable: true
        Hashes:
          items:
            $ref: '#/components/schemas/HashFilter'
          type: array
          nullable: true
        Path:
          $ref: '#/components/schemas/PathFilter'
        Signature:
          $ref: '#/components/schemas/SignatureFilter'
        WorkingDirectory:
          $ref: '#/components/schemas/PathFilter'
      required:
      - Path
    Assignment:
      example:
        Users:
        - DomainSid: DomainSid
          DomainName: DomainName
          AccountSid: AccountSid
          AccountName: AccountName
        - DomainSid: DomainSid
          DomainName: DomainName
          AccountSid: AccountSid
          AccountName: AccountName
        Profile:
          PromptSecureDesktop: true
          DefaultElevationKind: AutoApprove
          Id: Id
          ElevationSettings:
            Temporary:
              MaximumSeconds: 0
              Enabled: true
            Session:
              Enabled: true
          ElevationMethod: LocalAdmin
          Rules:
          - null
          - null
          Name: Unnamed profile
      properties:
        Profile:
          $ref: '#/components/schemas/Profile'
        Users:
          items:
            $ref: '#/components/schemas/User'
          type: array
      required:
      - Profile
      - Users
...
    ElevationConfigurations:
      example:
        Temporary:
          MaximumSeconds: 0
          Enabled: true
        Session:
          Enabled: true
      properties:
        Session:
          $ref: '#/components/schemas/SessionElevationConfiguration'
        Temporary:
          $ref: '#/components/schemas/TemporaryElevationConfiguration'
      required:
      - Session
      - Temporary
    ElevationKind:
      enum:
      - AutoApprove
      - Confirm
      - ReasonApproval
      - Deny
      type: string
    ElevationMethod:
      enum:
      - LocalAdmin
      - VirtualAccount
      type: string
...
    Error:
      enum:
      - AccessDenied
      - NotFound
      - InvalidParameter
      - Internal
      - Cancelled
      type: string
    ErrorResponse:
      example:
        Kind: AccessDenied
        Win32Error: 0
      properties:
        Kind:
          $ref: '#/components/schemas/Error'
        Win32Error:
          format: int32
          minimum: 0.0
          type: integer
      required:
      - Kind
      - Win32Error
    GetProfilesMeResponse:
      example:
        Active: Active
        Available:
        - null
        - null
      properties:
        Active:
          type: string
        Available:
          items:
            $ref: '#/components/schemas/Id'
          type: array
      required:
      - Available
...
    HashFilter:
      example:
        Sha1: Sha1
        Sha256: Sha256
      properties:
        Sha1:
          type: string
          nullable: true
        Sha256:
          type: string
          nullable: true
    Id:
      type: string
...
    OptionalId:
      example:
        Id: Id
      properties:
        Id:
          type: string
    PathFilter:
      example:
        Kind: Equals
        Data: Data
      properties:
        Data:
          type: string
        Kind:
          $ref: '#/components/schemas/PathFilterKind'
      required:
      - Data
      - Kind
    PathFilterKind:
      enum:
      - Equals
      - FileName
      - Wildcard
      type: string
...
    Profile:
      example:
        PromptSecureDesktop: true
        DefaultElevationKind: AutoApprove
        Id: Id
        ElevationSettings:
          Temporary:
            MaximumSeconds: 0
            Enabled: true
          Session:
            Enabled: true
        ElevationMethod: LocalAdmin
        Rules:
        - null
        - null
        Name: Unnamed profile
      properties:
        DefaultElevationKind:
          $ref: '#/components/schemas/ElevationKind'
        ElevationMethod:
          $ref: '#/components/schemas/ElevationMethod'
        ElevationSettings:
          $ref: '#/components/schemas/ElevationConfigurations'
        Id:
          type: string
        Name:
          default: Unnamed profile
          type: string
        PromptSecureDesktop:
          default: true
          type: boolean
        Rules:
          default: []
          items:
            $ref: '#/components/schemas/Id'
          type: array
    Rule:
      example:
        Target:
          Path:
            Kind: Equals
            Data: Data
          WorkingDirectory:
            Kind: Equals
            Data: Data
          CommandLine:
          - Kind: Equals
            Data: Data
          - Kind: Equals
            Data: Data
          Hashes:
          - Sha1: Sha1
            Sha256: Sha256
          - Sha1: Sha1
            Sha256: Sha256
          Signature:
            CheckAuthenticode: true
        ElevationKind: AutoApprove
        Id: Id
        Asker:
          Path:
            Kind: Equals
            Data: Data
          WorkingDirectory:
            Kind: Equals
            Data: Data
          CommandLine:
          - Kind: Equals
            Data: Data
          - Kind: Equals
            Data: Data
          Hashes:
          - Sha1: Sha1
            Sha256: Sha256
          - Sha1: Sha1
            Sha256: Sha256
          Signature:
            CheckAuthenticode: true
        Name: Name
      properties:
        Asker:
          $ref: '#/components/schemas/ApplicationFilter'
        ElevationKind:
          $ref: '#/components/schemas/ElevationKind'
        Id:
          type: string
        Name:
          type: string
        Target:
          $ref: '#/components/schemas/ApplicationFilter'
      required:
      - ElevationKind
      - Id
      - Name
      - Target
    SessionElevationConfiguration:
      example:
        Enabled: true
      properties:
        Enabled:
          type: boolean
      required:
      - Enabled
...
    SignatureFilter:
      example:
        CheckAuthenticode: true
      properties:
        CheckAuthenticode:
          type: boolean
      required:
      - CheckAuthenticode
...
    StringFilter:
      example:
        Kind: Equals
        Data: Data
      properties:
        Data:
          type: string
        Kind:
          $ref: '#/components/schemas/StringFilterKind'
      required:
      - Data
      - Kind
    StringFilterKind:
      enum:
      - Equals
      - Regex
      - StartsWith
      - EndsWith
      - Contains
      type: string
    TemporaryElevationConfiguration:
      example:
        MaximumSeconds: 0
        Enabled: true
      properties:
        Enabled:
          type: boolean
        MaximumSeconds:
          format: int64
          minimum: 0.0
          type: integer
      required:
      - Enabled
      - MaximumSeconds
...
    User:
      example:
        DomainSid: DomainSid
        DomainName: DomainName
        AccountSid: AccountSid
        AccountName: AccountName
      properties:
        AccountName:
          type: string
        AccountSid:
          type: string
        DomainName:
          type: string
        DomainSid:
          type: string
      required:
      - AccountName
      - AccountSid
      - DomainName
      - DomainSid